% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{extract_fit_parameters_sety}
\alias{extract_fit_parameters_sety}
\title{Title}
\usage{
extract_fit_parameters_sety(
  data,
  x_col = "Fraction_Old",
  y_col = "Pulse_time",
  id_cols = "variables",
  remove_cols = c("data", "model", "glance", "tidy")
)
}
\arguments{
\item{data}{tibble containing fitted model}

\item{x_col}{column containing fitted values}

\item{y_col}{}

\item{id_cols}{}

\item{remove_cols}{columns to remove after mutation}
}
\description{
Title
}
