% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{extract_fit_residuals}
\alias{extract_fit_residuals}
\title{Title}
\usage{
extract_fit_residuals(data, model_col = "model")
}
\arguments{
\item{data}{tibble of fitted models}

\item{model_col}{column name containing fitted models}
}
\description{
Title
}
