% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_density_modes}
\alias{find_density_modes}
\title{Title}
\usage{
find_density_modes(data, min_height = 0, topN = NULL, plot = F)
}
\arguments{
\item{data}{}

\item{min_height}{}

\item{topN}{}

\item{plot}{}
}
\description{
Title
}
