% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_lower_quantile}
\alias{in_lower_quantile}
\title{Title}
\usage{
in_lower_quantile(x, p = 0.99, na.rm = T)
}
\arguments{
\item{x}{numeric vector}

\item{p}{fraction of values to retain}
}
\description{
Title
}
