% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_n_sd}
\alias{in_n_sd}
\title{Title}
\usage{
in_n_sd(x, n = 3, center = "mean", na.rm = T)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of standard deviations}

\item{center}{method to determine center of distribution}
}
\description{
Title
}
