% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_SILAC_ratio}
\alias{is_SILAC_ratio}
\title{Test if element is a number between 0 and 1}
\usage{
is_SILAC_ratio(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Test if element is a number between 0 and 1
}
