% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_replace_all_m}
\alias{str_replace_all_m}
\title{Title}
\usage{
str_replace_all_m(sequence, mods = c(`SILAC-L` = "", `SILAC-H` = "UniMod:188"))
}
\arguments{
\item{sequence}{character vector}

\item{mods}{named vector c("pattern" = "replacement")}
}
\description{
Title
}
