% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_consecutive_timepoints}
\alias{filter_consecutive_timepoints}
\title{Title}
\usage{
filter_consecutive_timepoints(
  data,
  min_consecutive_timepoints = 2,
  only_keep_consecutive_timepoints = T,
  return_all_rows = T,
  value_col = "Fraction_Old",
  variables_col = "variables",
  observations_col = "observations",
  pulse_time_col = "Pulse_time",
  group_cols = c()
)
}
\arguments{
\item{data}{}

\item{min_consecutive_timepoints}{}

\item{only_keep_consecutive_timepoints}{}

\item{return_all_rows}{}

\item{value_col}{}

\item{variables_col}{}

\item{observations_col}{}

\item{pulse_time_col}{}

\item{group_cols}{}

\item{min_values_per_timepoint}{}

\item{only_keep_complete_timepoints}{}
}
\description{
Title
}
