% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_SILAC_plateaus_0}
\alias{filter_SILAC_plateaus_0}
\title{Title}
\usage{
filter_SILAC_plateaus_0(
  data,
  S2N_threshold = 2,
  return_all_rows = T,
  value_col = "Fraction_Old",
  variables_col = "variables",
  observations_col = "observations",
  pulse_time_col = "Pulse_time",
  group_cols = c()
)
}
\arguments{
\item{data}{}

\item{S2N_threshold}{}

\item{return_all_rows}{}

\item{value_col}{}

\item{variables_col}{}

\item{observations_col}{}

\item{pulse_time_col}{}

\item{group_cols}{}
}
\description{
Title
}
